Reworked file structure by seperating it. There was no logic implemented just better file structure.
