+++ src/common/constants.rs
@@ -1,0 +1,4 @@
+// Constants for API endpoints and defaults
+
+/// Default URL for the Ollama API
+pub const DEFAULT_URL: &str = "http://localhost:11434/api/chat";
