+++ src/common/mod.rs
@@ -1,0 +1,8 @@
+// Common module for simple_llama
+// Contains constants and utility functions
+
+mod constants;
+mod utils;
+
+// Re-export constants and utilities
+pub use constants::*;
