+++ src/common/utils.rs
@@ -1,0 +1,3 @@
+// Utility functions for simple_llama
+
+// Add any general utility functions here
