+++ src/lib.rs
@@ -1,99 +1,16 @@
-pub mod types;
-mod utils;
-use reqwest::Client;
-use types::{ChatMessage, Message, ModelMemory, ModelOptions};
+// Simple Llama - A Rust library for interacting with LLM models
+// This is the top-level module that re-exports everything
 
-/// Adds message to history vector
-/// 
-/// # Examples
-/// ```rust
-/// use simple_llama::{
-///     add_message,
-///     types::{ChatMessage, ModelMemory},
-/// };
-/// 
-/// let mut messages: ModelMemory = Vec::new();
-/// add_message(
-///     &mut messages,
-///     "system".to_string(),
-///     "Testing...".to_string(),
-/// );
-/// assert!(
-///     messages
-///         == [ChatMessage {
-///             role: "system".to_string(),
-///             content: "Testing...".to_string(),
-///         }]
-/// );
-/// ```
-pub fn add_message(model_memory: &mut ModelMemory, role: String, content: String) {
-    let msg = ChatMessage { role, content };
-    model_memory.push(msg);
-}
+// Module declarations
+pub mod chat;
+pub mod models;
+pub mod common;
 
-/// Sends message to ollama.
-/// 
-/// # Examples
-/// ```rust
-/// use reqwest::Client;
-/// use simple_llama::{
-///     send_message,
-///     types::{ModelOptions, ModelMemory, DEFAULT_URL, MessageMethods},
-/// };
-/// 
-/// #[tokio::main]
-/// async fn main() -> Result<(), Box<dyn std::error::Error>> {
-///     let client = Client::new();
-///     let data = ModelOptions {
-///         messages: ModelMemory::new(),
-///         temperature: 1.0,
-///         top_p: 1.0,
-///         top_k: 1,
-///         model: "llama2".to_string(),
-///         stream: false,
-///     };
-///     
-///     let response = send_message(&client, &data, DEFAULT_URL).await?;
-///     println!("Response: {}", response.get_llm_content());
-///     Ok(())
-/// }
-/// ```
-pub async fn send_message(
-    client: &Client,
-    data: &ModelOptions,
-    url: &str,
-) -> Result<Message, reqwest::Error> {
-    let result = client.post(url).json(&data).send().await?;
-    let status = result.status();
-    let text = result.text().await?;
+// Convenience re-exports
+pub use chat::{ChatMessage, Message, ModelMemory, add_message};
+pub use models::{ModelOptions, ModelData, send_message};
+pub use common::DEFAULT_URL;
 
-    Ok(Message {
-        status_code: status,
-        response: text,
-    })
-}
-
+// Legacy tests
 #[cfg(test)]
-mod tests {
-    use crate::{
-        add_message,
-        types::{ChatMessage, ModelMemory},
-    };
-
-    #[test]
-    fn test_add_message() {
-        let mut messages: ModelMemory = Vec::new();
-        add_message(
-            &mut messages,
-            "system".to_string(),
-            "Testing...".to_string(),
-        );
-        assert!(
-            messages
-                == [ChatMessage {
-                    role: "system".to_string(),
-                    content: "Testing...".to_string(),
-                }]
-        )
-    }
-}
+mod tests;
