--- src/utils.rs
@@ -1,2 +1,0 @@
-// utilities for lib.rs
-
